Update videoController.js - protect modifying videos

+++ src/controllers/videoController.js
@@ -31,22 +31,38 @@ export const watch = async (req, res) => {
 };
 
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    session: {
+      user: { _id },
+    },
+    params: { id },
+  } = req;
+  const video = await Video.findById(id);
+
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(404).redirect("/");
+  }
+  return res.render("edit", { pageTitle: `Editing: ${video.title}`, video });
+};
+
+export const postEdit = async (req, res) => {
+  const {
+    params: { id },
+    body: { title, description, hashtags },
+    session: {
+      user: { _id },
+    },
+  } = req;
   const video = await Video.findById(id);
 
   if (!video) {
     return res.render("404", { pageTitle: "Video not found." });
   }
-
-  return res.render("edit", { pageTitle: `Editing: ${video.title}`, video });
-};
-
-export const postEdit = async (req, res) => {
-  const { id } = req.params;
-  const video = await Video.findById(id);
-  const { title, description, hashtags } = req.body;
-  if (!video) {
-    return res.render("404", { pageTitle: "Video not found." });
+  if (String(video.owner) !== String(_id)) {
+    return res.status(404).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title: title,
@@ -92,7 +108,18 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: { user: _id },
+  } = req;
+  const video = await Video.findById(id);
+
+  if (!video) {
+    return res.render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(404).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
